+++ CHANGELOG.md
@@ -16,6 +16,7 @@ As this project is a user-facing application, the places in the semantic version
 ### Added
 
 - GRIDH logo in footer
+- In the year filter, click a histogram bar to select that decade (or century)
 
 ### Changed